Better handling of private pages

+++ src/Character.php
@@ -122,11 +122,13 @@ class Character extends AbstractTrackerEntity
             return 'Failed to get all necessary data for Character '.$this->id;
         }
         #Check if the character is private
+        $private = false;
         if (!empty($data['characters'][$this->id]) && is_array($data['characters'][$this->id]) && \array_key_exists('private', $data['characters'][$this->id]) && $data['characters'][$this->id]['private'] === true) {
             $this->markPrivate();
+            $private = true;
         }
         #Check for possible errors
-        if (empty($data['characters'][$this->id]['server'])) {
+        if (!$private && empty($data['characters'][$this->id]['server'])) {
             if (!empty($data['characters'][$this->id]) && (int)$data['characters'][$this->id] === 404) {
                 $this->delete();
                 return ['404' => true];
